Clean up pytest fixtures and helper functions
